Moved the functions for duplicating the ApplicationForm entries into `ApplicationFormTables`
