+++ src/main/scala/rifs/business/tables/OpportunityTables.scala
@@ -5,7 +5,7 @@ import javax.inject.Inject
 import org.joda.time.DateTime
 import play.api.db.slick.DatabaseConfigProvider
 import rifs.business.data.OpportunityOps
-import rifs.business.models.{ApplicationFormId, ApplicationFormSectionId, OpportunityId, OpportunityRow}
+import rifs.business.models.{OpportunityId, OpportunityRow}
 import rifs.business.restmodels.{Opportunity, OpportunitySummary, OpportunityValue}
 import rifs.business.slicks.modules.{ApplicationFormModule, OpportunityModule, PgSupport}
 import rifs.business.slicks.support.DBBinding
@@ -13,9 +13,8 @@ import slick.dbio.DBIOAction
 
 import scala.concurrent.{ExecutionContext, Future}
 
-class OpportunityTables @Inject()(val dbConfigProvider: DatabaseConfigProvider)(implicit ec: ExecutionContext)
+class OpportunityTables @Inject()(val dbConfigProvider: DatabaseConfigProvider, appFormTables: ApplicationFormTables)(implicit ec: ExecutionContext)
   extends OpportunityModule
-    with ApplicationFormModule
     with DBBinding
     with OpportunityOps
     with PgSupport {
@@ -51,7 +50,7 @@ class OpportunityTables @Inject()(val dbConfigProvider: DatabaseConfigProvider)(
           for {
             newId <- duplicateOpportunity(id, opp)
             _ <- duplicateOpportunitySections(id, newId)
-            _ <- duplicateApplicationForms(id, newId)
+            _ <- appFormTables.duplicateApplicationForms(id, newId)
           } yield Some(newId)
       }
     }
@@ -67,41 +66,6 @@ class OpportunityTables @Inject()(val dbConfigProvider: DatabaseConfigProvider)(
     sectionTable.filter(_.opportunityId === oldId).result.flatMap { sections =>
       (sectionTable ++= sections.map(_.copy(opportunityId = newId))).map(_ => ())
     }
-  }
-
-  private def duplicateApplicationForms(oldId: OpportunityId, newId: OpportunityId): DBIO[Unit] = {
-    val afIdio = for {
-      afs <- applicationFormTable.filter(_.opportunityId === oldId).result
-      newIds <- (applicationFormTable returning applicationFormTable.map(_.id)) ++= afs.map(_.copy(opportunityId = newId))
-    } yield afs.map(_.id).zip(newIds)
-
-    afIdio.flatMap { afIds =>
-      DBIO.sequence {
-        afIds.map { case (o, n) => duplicateAppFormSections(o, n) }
-      }.map(_ => ())
-    }
-  }
-
-  private def duplicateAppFormSections(oldId: ApplicationFormId, newId: ApplicationFormId): DBIO[Unit] = {
-    val afsIdio = for {
-      afss <- applicationFormSectionTable.filter(_.applicationFormId === oldId).result
-      newIds <- (applicationFormSectionTable returning applicationFormSectionTable.map(_.id)) ++= afss.map(_.copy(applicationFormId = newId))
-    } yield afss.map(_.id).zip(newIds)
-
-    afsIdio.flatMap { afsIds =>
-      DBIO.sequence {
-        afsIds.map { case (o, n) => duplicateAppFormQuestions(o, n) }
-      }.map(_ => ())
-    }
-  }
-
-  private def duplicateAppFormQuestions(oldId: ApplicationFormSectionId, newId: ApplicationFormSectionId): DBIO[Unit] = {
-    val afqIdio = for {
-      afqs <- applicationFormQuestionTable.filter(_.applicationFormSectionId === oldId).result
-      newIds <- (applicationFormQuestionTable returning applicationFormQuestionTable.map(_.id)) ++= afqs.map(_.copy(applicationFormSectionId = newId))
-    } yield afqs.map(_.id).zip(newIds)
-
-    afqIdio.map(_ => ())
   }
 
   /** ****************************
